Move all pagewriter related code to dedicated pagewriter package
