+++ pkg/app/pagewriter/error_page_test.go
@@ -1,4 +1,4 @@
-package app
+package pagewriter
 
 import (
 	"errors"
