+++ pkg/app/pagewriter/pagewriter_suite_test.go
@@ -1,4 +1,4 @@
-package app
+package pagewriter
 
 import (
 	"testing"
