+++ pkg/app/pagewriter/sign_in_page_test.go
@@ -1,4 +1,4 @@
-package app
+package pagewriter
 
 import (
 	"html/template"
